ACADEMe-backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
passlib[bcrypt]
pyjwt
pydantic
pytest
httpx
pydantic[email]
google-generativeai
cloudinary
python-multipart
